@@ -12,65 +16,81 @@
     "install-on-device:prod": "adb install -r ./android/app/build/outputs/apk/prod/release/app-prod-release.apk",
     "lint": "eslint ./app/**/*.js",
     "lint:fix": "yarn lint --fix",
-    "run:android": "react-native run-android --variant=devDebug",
     "start": "react-native start",
-    "test": "jest"
+    "start:android": "react-native run-android --variant=devDebug",
+    "test": "jest",
+    "postinstall": "npx jetify"
   },
   "dependencies": {
-    "i18n-js": "3.2.1",
-    "jsc-android": "^236355.1.1",
-    "popcorn-sdk": "^3.1.0",
+    "@apollo/react-common": "3.1.2",
+    "@apollo/react-components": "3.1.2",
+    "@apollo/react-hooks": "3.1.2",
+    "@react-native-community/async-storage": "1.6.2",
+    "@react-native-community/netinfo": "^4.4.0",
+    "apollo-cache-inmemory": "^1.6.3",
+    "apollo-cache-persist": "^0.1.1",
+    "apollo-client": "^2.6.4",
+    "apollo-link": "^1.2.13",
+    "apollo-link-http": "^1.5.16",
+    "apollo-link-ws": "^1.0.19",
+    "apollo-utilities": "^1.3.2",
+    "graphql": "14.5.8",
+    "graphql-tag": "^2.10.1",
+    "hoist-non-react-statics": "^3.3.0",
+    "i18n-js": "3.3.0",
+    "jsc-android": "241213.1.0",
+    "lottie-react-native": "^3.2.1",
     "prop-types": "15.7.2",
-    "react": "16.8.4",
-    "react-native": "0.59.1",
-    "react-native-animatable": "1.3.1",
-    "react-native-fs": "2.13.3",
-    "react-native-gesture-handler": "1.1.0",
-    "react-native-google-cast": "github:tripss/react-native-google-cast#8e71424503e5b86a443f0ff7ddbdfc4d72d3a754",
+    "react": "16.10.2",
+    "react-native": "0.61.2",
+    "react-native-animatable": "1.3.2",
+    "react-native-fs": "^2.14.1",
+    "react-native-gesture-handler": "1.4.1",
+    "react-native-google-cast": "^3.1.3",
     "react-native-languages": "3.0.2",
-    "react-native-linear-gradient": "2.5.4",
+    "react-native-linear-gradient": "2.5.6",
     "react-native-markdown-renderer": "^3.2.8",
+    "react-native-network-info": "^5.2.1",
     "react-native-orientation": "^3.1.3",
     "react-native-splash-screen": "3.2.0",
-    "react-native-static-server": "0.4.1",
-    "react-native-torrent-streamer": "github:tripss/react-native-torrent-streamer",
-    "react-native-typography": "^1.3.0",
-    "react-native-vector-icons": "6.4.1",
-    "react-native-video": "4.4.0",
-    "react-navigation": "3.4.1",
-    "react-redux": "6.0.1",
-    "redux": "^4.0.0",
-    "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.3.0",
+    "react-native-static-server": "0.4.2",
+    "react-native-vector-icons": "6.6.0",
+    "react-navigation": "4.0.10",
+    "react-navigation-stack": "1.9.4",
+    "redux-persist-fs-storage": "^1.3.0",
+    "subscriptions-transport-ws": "^0.9.16",
     "update-react-native-app": "1.1.2"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "7.3.4",
-    "@babel/plugin-proposal-decorators": "7.3.0",
-    "@babel/plugin-proposal-export-namespace-from": "^7.2.0",
-    "@babel/plugin-proposal-object-rest-spread": "7.3.4",
-    "@babel/plugin-transform-flow-strip-types": "7.3.4",
-    "@babel/plugin-transform-runtime": "7.3.4",
-    "@babel/preset-react": "^7.0.0",
-    "babel-eslint": "^10.0.1",
+    "@babel/plugin-proposal-class-properties": "7.5.5",
+    "@babel/plugin-proposal-decorators": "7.6.0",
+    "@babel/plugin-proposal-export-namespace-from": "7.5.2",
+    "@babel/plugin-proposal-object-rest-spread": "7.6.2",
+    "@babel/plugin-transform-flow-strip-types": "7.6.3",
+    "@babel/plugin-transform-runtime": "7.6.2",
+    "@babel/preset-react": "7.6.3",
+    "@commitlint/cli": "8.2.0",
+    "@commitlint/config-conventional": "8.2.0",
+    "babel-eslint": "10.0.3",
     "babel-plugin-module-resolver": "3.2.0",
-    "eslint": "5.15.3",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-config-standard-react": "^7.0.2",
+    "eslint": "6.5.1",
+    "eslint-config-airbnb": "18.0.1",
+    "eslint-config-standard": "14.1.0",
+    "eslint-config-standard-react": "9.2.0",
     "eslint-plugin-babel": "^5.2.1",
-    "eslint-plugin-import": "2.16.0",
-    "eslint-plugin-jest": "22.4.1",
-    "eslint-plugin-jsx-a11y": "6.2.1",
-    "eslint-plugin-node": "8.0.1",
-    "eslint-plugin-promise": "^4.0.1",
-    "eslint-plugin-react": "7.12.4",
-    "eslint-plugin-standard": "^4.0.0",
+    "eslint-plugin-import": "2.18.2",
+    "eslint-plugin-jest": "22.17.0",
+    "eslint-plugin-jsx-a11y": "6.2.3",
+    "eslint-plugin-node": "10.0.0",
+    "eslint-plugin-promise": "4.2.1",
+    "eslint-plugin-react": "7.16.0",
+    "eslint-plugin-standard": "4.0.1",
     "get-dev-paths": "^0.1.1",
-    "jest": "24.5.0",
+    "husky": "3.0.8",
+    "jest": "24.9.0",
     "jest-react-native": "^18.0.0",
-    "metro-react-native-babel-preset": "0.53.1",
-    "react-test-renderer": "16.8.4"
+    "metro-react-native-babel-preset": "0.56.0",
+    "react-test-renderer": "16.10.2"
   },
   "jest": {
     "preset": "react-native"
