@@ -8,10 +8,13 @@ const styles = StyleSheet.create({
 
   image: {
     height: '100%',
-    width : '100%',
+    width: '100%',
   },
 
 })
+
+// https://github.com/DylanVann/react-native-fast-image
+// https://github.com/DylanVann/react-native-fast-image#sourcecache-enum
 
 export default class Image extends React.PureComponent {
 
@@ -25,23 +28,24 @@ export default class Image extends React.PureComponent {
 
   static propTypes = {
     images: PropTypes.object.isRequired,
-    type  : PropTypes.oneOf([
+    type: PropTypes.oneOf([
       'poster',
-      'fanart',
+      'backdrop',
     ]),
-    size  : PropTypes.string,
-
-    style       : PropTypes.object,
+    size: PropTypes.string,
+    style: PropTypes.object,
     withFallback: PropTypes.bool,
-    resizeMode  : PropTypes.string,
+    empty: PropTypes.bool,
+    resizeMode: PropTypes.string,
   }
 
   static defaultProps = {
-    type        : 'poster',
-    size        : 'thumb',
-    resizeMode  : 'stretch',
-    style       : null,
+    type: 'poster',
+    size: 'thumb',
+    resizeMode: 'stretch',
+    style: null,
     withFallback: true,
+    empty: false,
   }
 
   state = {
