@@ -13,38 +13,39 @@ export default class Typography extends Component {
 
     variant: PropTypes.oneOf([
       'default',
-      'display1',
-      'display2',
-      'display3',
-      'display4',
-      'display5',
-      'headline',
-      'title',
-      'subheading',
-      'body2',
+      'headline1',
+      'headline2',
+      'headline3',
+      'headline4',
+      'headline5',
+      'headline6',
+      'subtitle1',
+      'subtitle2',
       'body1',
-      'caption',
+      'body2',
       'button',
+      'caption',
+      'captionSmall',
+      'overline',
     ]),
 
     color: PropTypes.oneOf([
       'white',
       'black',
+      'primary',
+      'primaryDark',
     ]),
 
     emphasis: PropTypes.oneOf([
       'high',
       'medium',
+      'low',
     ]),
 
     fontWeight: PropTypes.oneOf([
-      'thin',
-      'extraLight',
-      'light',
+      'default',
       'regular',
       'medium',
-      'bold',
-      'black',
     ]),
 
     children: PropTypes.oneOfType([
@@ -62,14 +63,14 @@ export default class Typography extends Component {
   }
 
   static defaultProps = {
-    emphasis  : 'high',
-    color     : 'white',
-    variant   : 'default',
-    fontWeight: 'regular',
-    textProps : {},
-    component : null,
-    children  : null,
-    style     : null,
+    emphasis: 'high',
+    color: 'white',
+    variant: 'default',
+    fontWeight: 'default',
+    textProps: {},
+    component: null,
+    children: null,
+    style: null,
   }
 
   static getTextStyle = ({
@@ -81,10 +82,16 @@ export default class Typography extends Component {
     asObject = false,
   }) => {
     const styledColor = styles[`color${capitalizeFirstLetter(color)}`]
-    const styledFont = styles[`fontFamily${capitalizeFirstLetter(fontWeight)}`]
+    const styledFont = fontWeight !== 'default' ? styles[`fontFamily${capitalizeFirstLetter(fontWeight)}`] : null
     const styledEmphasis = styles[`emphasis${capitalizeFirstLetter(emphasis)}`]
 
-    const textStyles = [styledColor, styles[variant], styledFont, styleProp, styledEmphasis]
+    const textStyles = [
+      styledColor,
+      styles[variant],
+      styledFont,
+      styleProp,
+      color !== 'primary' || emphasis !== 'high' ? styledEmphasis : {},
+    ]
 
     if (asObject) {
       let objectStyle = {}
