@@ -1,4 +1,4 @@
-export const fontName = 'Roboto'
+export const fontName = 'Quicksand'
 
 import colors from 'modules/colors'
 
@@ -6,81 +6,105 @@ export default {
 
   root: {},
 
-  display5: {
-    fontSize  : 22,
+
+  headline1: {
+    fontFamily: `${fontName}-Regular`,
+    lineHeight: 122,
+    letterSpacing: -1.4,
+    fontSize: 96,
+  },
+
+  headline2: {
+    fontFamily: `${fontName}-Regular`,
+    lineHeight: 72,
+    letterSpacing: -0.5,
+    fontSize: 60,
+  },
+
+  headline3: {
+    fontFamily: `${fontName}-Regular`,
+    lineHeight: 56,
+    letterSpacing: 0,
+    fontSize: 48,
+  },
+
+  headline4: {
+    fontFamily: `${fontName}-Regular`,
+    lineHeight: 46,
+    letterSpacing: 0,
+    fontSize: 34,
+  },
+
+  headline5: {
+    fontFamily: `${fontName}-Regular`,
     lineHeight: 24,
-    letterSpacing: 0.25,
+    letterSpacing: 0.18,
+    fontSize: 24,
   },
 
-  display4: {
-    fontSize  : 26,
-    lineHeight: 30,
-    letterSpacing: 0.19,
-  },
-
-  display3: {
-    fontSize  : 56,
-    lineHeight: 64,
-  },
-
-  display2: {
-    fontSize  : 45,
-    lineHeight: 52,
-  },
-
-  display1: {
-    fontSize  : 34,
-    lineHeight: 40,
-  },
-
-  headline: {
-    fontSize  : 24,
-    lineHeight: 32,
-  },
-
-  title: {
-    fontSize  : 20,
+  headline6: {
+    fontFamily: `${fontName}-Medium`,
     lineHeight: 24,
-    letterSpacing: 0.25,
+    letterSpacing: 0.15,
+    fontSize: 20,
   },
 
-  subheading: {
-    fontSize  : 16,
+  subtitle1: {
+    fontFamily: `${fontName}-Regular`,
     lineHeight: 24,
+    letterSpacing: 0.15,
+    fontSize: 16,
   },
 
-  body2: {
-    fontSize  : 12,
-    lineHeight: 15,
-    letterSpacing: 0.21
+  subtitle2: {
+    fontFamily: `${fontName}-Medium`,
+    lineHeight: 24,
+    letterSpacing: 0.10,
+    fontSize: 14,
   },
 
   body1: {
-    fontSize  : 14,
-    lineHeight: 20,
+    fontFamily: `${fontName}-Regular`,
+    lineHeight: 24,
+    letterSpacing: 0.50,
+    fontSize: 16,
   },
 
-  caption: {
-    fontSize     : 12,
-    lineHeight   : 16,
-    letterSpacing: 0.4,
+  body2: {
+    fontFamily: `${fontName}-Regular`,
+    lineHeight: 20,
+    letterSpacing: 0.25,
+    fontSize: 14,
   },
 
   button: {
-    fontSize  : 14,
-    lineHeight: 20,
+    fontFamily: `${fontName}-Medium`,
+    lineHeight: 16,
+    letterSpacing: 0.10,
+    fontSize: 14,
+    textTransform: 'uppercase',
   },
 
-  fontFamilyThin: {
-    fontFamily: `${fontName}-Thin`,
+  caption: {
+    fontFamily: `${fontName}-Regular`,
+    lineHeight: 16,
+    letterSpacing: 0.40,
+    fontSize: 12,
   },
 
-  fontFamilyExtraLight: {
-    fontFamily: `${fontName}-ExtraLight`,
+  captionSmall: {
+    fontFamily: `${fontName}-Regular`,
+    lineHeight: 16,
+    letterSpacing: 0.40,
+    fontSize: 10,
   },
 
-  fontFamilyLight: {
-    fontFamily: `${fontName}-Light`,
+  overline: {
+    fontFamily: `${fontName}-Medium`,
+    lineHeight: 16,
+    letterSpacing: 1.50,
+    fontSize: 12,
+    textTransform: 'uppercase',
   },
 
   fontFamilyRegular: {
@@ -91,25 +115,32 @@ export default {
     fontFamily: `${fontName}-Medium`,
   },
 
-  fontFamilyBold: {
-    fontFamily: `${fontName}-Bold`,
+  colorPrimary: {
+    color: colors.PRIMARY_COLOR_200,
   },
 
-  fontFamilyBlack: {
-    fontFamily: `${fontName}-Black`,
+  colorPrimaryDark: {
+    color: colors.PRIMARY_COLOR_500,
   },
 
   colorWhite: {
-    color: colors.TEXT_COLOR,
+    color: '#ffffff',
   },
 
   colorBlack: {
     color: 'black',
   },
 
-  emphasisHigh: {},
+  emphasisHigh: {
+    opacity: 0.87,
+  },
 
   emphasisMedium: {
-    opacity: 0.6,
+    opacity: 0.60,
   },
+
+  emphasisLow: {
+    opacity: 0.38,
+  },
+
 }
