@@ -1,19 +1,19 @@
+const PRIMARY_COLOR_200 = '#FECC91'
+const PRIMARY_COLOR_500 = '#FB9B42'
+
+const BACKGROUND = '#202125'
+
 export default {
-  BACKGROUND        : '#000000',
-  BACKGROUND_DARKER : '#04090c',
-  BACKGROUND_LIGHTER: '#212121',
 
-  TAB_NAVIGATION: '#0D0E0F',
+  PRIMARY_COLOR_200,
+  PRIMARY_COLOR_500,
+  BACKGROUND,
 
-  BUTTON_DEFAULT: '#17181b',
-  BUTTON_PRIMARY: '#212121',
+  BACKGROUND_OVERLAY: '#000000',
 
-  TEXT_COLOR: '#FFFFFF',
-  ICON_COLOR: '#FFFFFF',
-
-  BOTTOM_BAR_INACTIVE_TINT: '#939597',
-  BOTTOM_BAR_ACTIVE_TINT  : '#FAFAFA',
-
-  ITEM_ACTIVE_TAB_INDICATOR: '#FAFAFA',
-  ITEM_DROPDOWN_BACKGROUND : '#212121',
+  ICON: {
+    WHITE: '#ffffff',
+    PRIMARY: PRIMARY_COLOR_200,
+    '01DP': '#2B2C30',
+  },
 }
